Synchronize database.update in UserController
The editTarget and editLeniencyLevel functions are called
at the same time from the front-end, resulting in the call
to the database needing to be synchronized to prevent a
concurrent modification exception.

May want to make all calls to database.update synchronized
in the future in case API endpoints are called differently
to prevent future errors.